Cleaned up StreamProcessor and WaveformWidget code.

+++ src/streamprocessor/streamprocessor.cpp
@@ -476,10 +476,10 @@ StreamProcessor::processAudio()
 
 					if(m_swResample) {
 						emit audioDataAvailable(frameResampled->data[0], frameSize * frameResampled->channels,
-							&m_audioStreamFormat, timeFrameStart + timeResampleDelay, timeFrameDuration);
+							&m_audioStreamFormat, qint64(timeFrameStart + timeResampleDelay), qint64(timeFrameDuration));
 					} else {
 						emit audioDataAvailable(frame->data[0], frameSize * frame->channels,
-							&m_audioStreamFormat, timeFrameStart, timeFrameDuration);
+							&m_audioStreamFormat, qint64(timeFrameStart), qint64(timeFrameDuration));
 					}
 
 					drainSampleBuffer = swr_get_out_samples(m_swResample, 0) > 1000;
